+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🐛 fix: messages on errors fixed
 * 🎉 feat: added a validator and refactor code for register logic
 * 🎉 feat: added config loading
 * 🎉 feat: added config loading